requirements.txt
aiohttp==3.8.1
aiomqtt==0.1.3
aiomysql==0.0.22
aioredis==2.0.1
aiosignal==1.2.0
aiosqlite==0.17.0
amqp==5.0.9
anyio==3.5.0
asgiref==3.4.1
async-timeout==4.0.2
asyncpg==0.25.0
attrs==21.4.0
bcrypt==3.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.11
click==8.0.3
contextlib2==21.6.0
contextvars==2.4
dataclasses==0.6
dateutils==0.6.12
Deprecated==1.2.13
dnspython==2.2.0
docutils==0.18.1
email-validator==1.1.3
fastapi==0.73.0
fastapi-users==9.2.4
fastapi-users-db-sqlalchemy==2.0.4
frozenlist==1.3.0
future==0.18.2
greenlet==1.1.2
gunicorn==20.1.0
h11==0.13.0
idna==3.3
immutables==0.16
importlib-metadata==4.8.3
jproperties==2.1.1
lockfile==0.12.2
makefun==1.13.1
multidict==6.0.2
multiprocessing-logging==0.3.1
mysql-connector-python==8.0.28
oauthlib==3.2.0
packaging==21.3
paho-mqtt==1.5.1
passlib==1.7.4
protobuf==3.19.4
psutil==5.9.0
py-postgresql==1.2.2
pycparser==2.21
pycrypto==2.6.1
pydantic==1.9.0
PyJWT==2.3.0
PyMySQL==0.9.3
pyparsing==3.0.7
python-daemon==2.3.0
python-dateutil==2.8.2
python-dotenv==0.19.2
python-multipart==0.0.5
pytz==2021.3
PyYAML==6.0
redis==4.1.3
requests==2.27.1
requests-jwt==0.6.0
requests-oauthlib==1.3.1
six==1.16.0
sniffio==1.2.0
solrpy==1.0.0
SQLAlchemy==1.4.31
starlette==0.17.1
stompest==2.3.0
textwrap3==0.9.2
typing_extensions==4.0.1
urllib3==1.26.8
uvicorn==0.16.0
vine==5.0.0
wrapt==1.13.3
yarl==1.7.2
zipp==3.6.0
